RPMB: solving the problem for creating two different set of TEE images
In the current implementation of RPMB in OPTEE there is a flaw like
two different set of TEE binary need to be created: factory &
production to prevent the attack. While the production image will
not be useful if one replace the emmc from the device.

This patch deal with this problem. In the key generation process,
in a fresh target with new emmc while doing the SHA-256-HMAC, a random
number is created and updated to form the final RPMB key. The key is
being written to RPMB OTP, and the random number generated during
the process is also written to a fixed partition area.

In this Proof-of-concept, RPMB address 256 (block index 1) is being
used. Please note that this patch don't implement any check on RPMB
write on block index 1, which need to be implemented for a safer
side.

The changes has been unit tested with tee-supplicant RPMB-EMU.

Signed-off-by: Sourabh Das <sourabhemsec@outlook.com>